refactor: extract AppState into its own module
- Move AppState, its impls, and FromRef wirings out of lib.rs into
  src/state.rs; lib.rs re-exports AppState.
- Collapse panes::activate handler into a thin shim over the new
  AppState::activate_pane method (mirrors activities::activate).
