+++ daemon/http_server/src/state.rs
@@ -1,0 +1,157 @@
+use crate::HttpResult;
+use crate::handlers::windows;
+use crate::layout_broadcast::LayoutBroadcaster;
+use axum::extract::FromRef;
+use ozmux_extension::ExtensionRegistry;
+use ozmux_multiplexer::{
+    Activity, ActivityId, MultiplexerError, MultiplexerResult, MultiplexerService, PaneId,
+    SessionId, SetActiveOutcome, SetActivePaneOutcome, WindowId,
+};
+use ozmux_terminal::TerminalService;
+
+#[derive(Clone)]
+pub struct AppState {
+    pub multiplexer: MultiplexerService,
+    pub terminal: TerminalService,
+    pub extensions: ExtensionRegistry,
+    pub layout_broadcast: LayoutBroadcaster,
+}
+
+impl AppState {
+    /// Build an `AppState` wired to the supplied runtime services. This is the
+    /// only sanctioned construction path outside tests — `Default` is
+    /// intentionally not derived so callers cannot silently produce a state
+    /// whose `TerminalService`, `ExtensionRegistry`, or `LayoutBroadcaster`
+    /// are detached from the daemon's runtime root.
+    pub fn new(
+        terminal: TerminalService,
+        extensions: ExtensionRegistry,
+        layout_broadcast: LayoutBroadcaster,
+    ) -> Self {
+        Self {
+            multiplexer: MultiplexerService::default(),
+            terminal,
+            extensions,
+            layout_broadcast,
+        }
+    }
+
+    pub async fn activate_activity(
+        &self,
+        wid: &WindowId,
+        pid: &PaneId,
+        aid: &ActivityId,
+    ) -> HttpResult {
+        let outcome = self
+            .multiplexer
+            .with_window_or_404(wid, |w| w.pane_mut(pid)?.set_active_activity(aid))
+            .await?;
+        if matches!(outcome, SetActiveOutcome::Changed) {
+            self.publish_window_layout(wid).await;
+        }
+        Ok(())
+    }
+
+    pub async fn activate_pane(&self, wid: &WindowId, pid: &PaneId) -> HttpResult {
+        let outcome = self
+            .multiplexer
+            .with_window_or_404(wid, |w| {
+                if w.panes.contains_key(pid) {
+                    w.set_active_pane(pid)
+                } else if self.multiplexer.pane_owner_window.contains_key(pid) {
+                    Err(MultiplexerError::PaneNotInWindow {
+                        window: w.id.clone(),
+                        pane: pid.clone(),
+                    })
+                } else {
+                    Err(MultiplexerError::PaneNotFound(pid.clone()))
+                }
+            })
+            .await?;
+        if matches!(outcome, SetActivePaneOutcome::Changed) {
+            self.publish_window_layout(wid).await;
+        }
+        Ok(())
+    }
+
+    pub async fn add_activity_to_pane(
+        &self,
+        wid: &WindowId,
+        pid: &PaneId,
+        activity: Activity,
+        extension_name: Option<&str>,
+    ) -> HttpResult<ActivityId> {
+        let aid = activity.id.clone();
+        self.multiplexer
+            .with_window_or_404(wid, |w| w.pane_mut(pid)?.add_activity(activity))
+            .await?;
+        if let Some(name) = extension_name {
+            self.extensions.record_activity_owner(&aid, name);
+        }
+        self.publish_window_layout(wid).await;
+        Ok(aid)
+    }
+
+    /// Build the current Window layout snapshot under the Window lock and
+    /// broadcast it. Used by every handler that mutates a Window.
+    async fn publish_window_layout(&self, wid: &WindowId) {
+        let _ = self
+            .multiplexer
+            .with_window(wid, |w| match windows::window_view_for(w) {
+                Ok(view) => self.layout_broadcast.publish(wid, view),
+                Err(e) => tracing::warn!(error = %e, %wid, "skipped layout publish"),
+            })
+            .await;
+    }
+
+    /// Close a Window: tear down its Panes/Activities and run runtime
+    /// cleanup. Delegates the data half to `multiplexer.close_window_data`
+    /// and then issues PTY kills, extension registry forgets, and a layout
+    /// broadcast close.
+    pub async fn close_window(&self, wid: &WindowId) -> MultiplexerResult<Vec<ActivityId>> {
+        let (activities, pane_ids) = self.multiplexer.close_window_data(wid).await?;
+        for pid in pane_ids {
+            self.extensions.forget_pane(&pid);
+        }
+        for aid in &activities {
+            let _ = self.terminal.kill(aid).await;
+            self.extensions.forget_activity(aid);
+        }
+        self.layout_broadcast.close(wid);
+        Ok(activities)
+    }
+
+    /// Delete a Session, cascading into every Window it owns.
+    pub async fn delete_session(&self, sid: &SessionId) -> MultiplexerResult<Vec<ActivityId>> {
+        let linked = self.multiplexer.take_session_windows(sid).await?;
+        let mut activities = Vec::new();
+        for wid in linked {
+            activities.extend(self.close_window(&wid).await?);
+        }
+        Ok(activities)
+    }
+}
+
+impl FromRef<AppState> for TerminalService {
+    fn from_ref(input: &AppState) -> Self {
+        input.terminal.clone()
+    }
+}
+
+impl FromRef<AppState> for ExtensionRegistry {
+    fn from_ref(input: &AppState) -> Self {
+        input.extensions.clone()
+    }
+}
+
+impl FromRef<AppState> for LayoutBroadcaster {
+    fn from_ref(input: &AppState) -> Self {
+        input.layout_broadcast.clone()
+    }
+}
+
+impl FromRef<AppState> for MultiplexerService {
+    fn from_ref(input: &AppState) -> Self {
+        input.multiplexer.clone()
+    }
+}
